requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
autopy==4.0.0
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets==2.14.4
dill==0.3.7
distro==1.9.0
dlib @ file:///D:/bld/dlib-split_1716026478642/work
einops==0.6.1
filelock==3.9.0
fonttools==4.39.3
frozenlist==1.4.0
fsspec==2023.5.0
h5py==3.11.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-resources @ file:///C:/b/abs_9d6rmtszbl/croot/importlib_resources_1698254653090/work
imutils==0.5.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
lazy_loader==0.2
Levenshtein==0.21.1
lightly==1.4.12
lightly-utils==0.0.2
lightning-utilities==0.9.0
lpips==0.1.4
MarkupSafe==2.1.2
matplotlib @ file:///C:/b/abs_085jhivdha/croot/matplotlib-suite_1693812524572/work
metrics==0.3.3
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
munkres==1.1.4
natsort==8.3.1
networkx==3.1
nltk==3.8.1
nougat-ocr==0.1.2
numexpr @ file:///C:/b/abs_afm0oewmmt/croot/numexpr_1683221839116/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
omegaconf==2.3.0
opencv-contrib-python==4.5.5.62
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.62
orjson==3.9.5
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1692213983661/work
pandoc==2.3
pathlib2==2.3.7.post1
pathspec==0.5.5
patsy==0.5.6
pillow @ file:///D:/bld/pillow_1704252195854/work
plumbum==1.8.1
ply==3.11
psutil==6.0.0
pyarrow==13.0.0
pydantic==1.10.11
pygame==2.5.2
Pygments==2.2.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
pyreadline3==3.4.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-Levenshtein==0.21.1
pytorch-lightning==2.0.5
pytz @ file:///C:/b/abs_19q3ljkez4/croot/pytz_1695131651401/work
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
qudida==0.0.4
rapidfuzz==3.2.0
regex==2023.8.8
requests==2.28.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
scikit-build==0.17.6
scikit-image==0.18.0
scikit-learn==1.2.2
scipy==1.9.1
sconf==0.2.5
seaborn @ file:///C:/b/abs_68ltdkoyoo/croot/seaborn_1673479199997/work
sentencepiece==0.1.99
shapely==2.0.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post4
statsmodels==0.14.1
sympy==1.11.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.5.4
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0+cu117
torch_geometric==2.5.3
torchaudio==2.0.1+cu117
torchdata==0.8.0
torchinfo==1.8.0
torchmetrics==1.0.1
torchsummary==1.5.1
torchvision==0.15.1+cu117
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.4.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==1.26.13
xxhash==3.3.0
yarl==1.9.2
zipp==3.15.0
